pyproject.toml
[project]
name = "til-migration-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.4",
    "langchain-google-genai>=4.1.2",
    "langchainhub>=0.1.21",
    "langsmith>=0.5.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
]
